face-rekon/requirements-test.txt
flask==3.1.2
flask-cors==6.0.1
flask-restx==1.3.0
pillow==11.3.0
numpy==2.2.6
pytest==7.4.0
pytest-cov==4.1.0